ref(ts): Convert `<TimeSince>` to typescript
typescript me
